fix workspaces tags
