+++ aws/internal/keyvaluetags/service_generation_customizations.go
@@ -576,6 +576,8 @@ func ServiceTagFunction(serviceName string) string {
 		return "AddTagsToResource"
 	case "storagegateway":
 		return "AddTagsToResource"
+	case "workspaces":
+		return "CreateTags"
 	default:
 		return "TagResource"
 	}
@@ -696,6 +698,8 @@ func ServiceTagInputIdentifierField(serviceName string) string {
 		return "ResourceARN"
 	case "wafv2":
 		return "ResourceARN"
+	case "workspaces":
+		return "ResourceId"
 	default:
 		return "ResourceArn"
 	}
@@ -881,6 +885,8 @@ func ServiceUntagFunction(serviceName string) string {
 		return "RemoveTagsFromResource"
 	case "storagegateway":
 		return "RemoveTagsFromResource"
+	case "workspaces":
+		return "DeleteTags"
 	default:
 		return "UntagResource"
 	}
