+++ aws/internal/keyvaluetags/update_tags_gen.go
@@ -105,6 +105,7 @@ import (
 	"github.com/aws/aws-sdk-go/service/waf"
 	"github.com/aws/aws-sdk-go/service/wafregional"
 	"github.com/aws/aws-sdk-go/service/wafv2"
+	"github.com/aws/aws-sdk-go/service/workspaces"
 )
 
 // AccessanalyzerUpdateTags updates accessanalyzer service tags.
@@ -3667,6 +3668,42 @@ func Wafv2UpdateTags(conn *wafv2.WAFV2, identifier string, oldTagsMap interface{
 		}
 
 		_, err := conn.TagResource(input)
+
+		if err != nil {
+			return fmt.Errorf("error tagging resource (%s): %w", identifier, err)
+		}
+	}
+
+	return nil
+}
+
+// WorkspacesUpdateTags updates workspaces service tags.
+// The identifier is typically the Amazon Resource Name (ARN), although
+// it may also be a different identifier depending on the service.
+func WorkspacesUpdateTags(conn *workspaces.WorkSpaces, identifier string, oldTagsMap interface{}, newTagsMap interface{}) error {
+	oldTags := New(oldTagsMap)
+	newTags := New(newTagsMap)
+
+	if removedTags := oldTags.Removed(newTags); len(removedTags) > 0 {
+		input := &workspaces.DeleteTagsInput{
+			ResourceId: aws.String(identifier),
+			TagKeys:    aws.StringSlice(removedTags.IgnoreAws().Keys()),
+		}
+
+		_, err := conn.DeleteTags(input)
+
+		if err != nil {
+			return fmt.Errorf("error untagging resource (%s): %w", identifier, err)
+		}
+	}
+
+	if updatedTags := oldTags.Updated(newTags); len(updatedTags) > 0 {
+		input := &workspaces.CreateTagsInput{
+			ResourceId: aws.String(identifier),
+			Tags:       updatedTags.IgnoreAws().WorkspacesTags(),
+		}
+
+		_, err := conn.CreateTags(input)
 
 		if err != nil {
 			return fmt.Errorf("error tagging resource (%s): %w", identifier, err)
